+++ src/components/SegmentedControl/index.ts
@@ -3,4 +3,5 @@ export type {
   SegmentedControlProps,
   SegmentedOption,
   SegmentedControlColorProps,
+  SegmentedControlScrollButtonsProps,
 } from './SegmentedControl.types';