+++ src/components/Select/Select.component.tsx
@@ -3,7 +3,7 @@ import { X, Check } from 'lucide-react';
 import { Input } from '../Input/Input.component';
 import { Dropdown } from '../Dropdown/Dropdown.component';
 import type { SelectProps, SelectOption } from './Select.types';
-import { renderIcon } from '../../utils';
+import { renderIcon, fullWidthModifier } from '../../utils';
 
 export const Select = React.forwardRef<HTMLDivElement, SelectProps>(
   (
@@ -325,7 +325,7 @@ export const Select = React.forwardRef<HTMLDivElement, SelectProps>(
     const triggerElement = (
       <div
         ref={triggerRef}
-        className={`eidos-select-trigger ${fullWidth ? 'eidos-select-trigger--fullWidth' : ''}`}
+        className={`eidos-select-trigger ${fullWidth ? fullWidthModifier('eidos-select-trigger') : ''}`}
         onKeyDown={handleKeyDown}
         tabIndex={disabled ? -1 : undefined}
         role="combobox"
@@ -359,7 +359,7 @@ export const Select = React.forwardRef<HTMLDivElement, SelectProps>(
       <div
         ref={ref}
         className={`${'eidos-select-container'} ${
-          fullWidth ? 'eidos-select-container--fullWidth' : ''
+          fullWidth ? fullWidthModifier('eidos-select-container') : ''
         } ${className}`}
       >
         <Dropdown